grpc access logger: refactor common gRPC ALS out (#7879)
Description:
In preparation to implement TCP gRPC Access Logger.

Risk Level: Low (refactoring only)
Testing: CI
Docs Changes: N/A
Release Notes: N/A
