+++ source/extensions/access_loggers/grpc/grpc_access_log_impl.h
@@ -19,7 +19,7 @@
 namespace Envoy {
 namespace Extensions {
 namespace AccessLoggers {
-namespace HttpGrpc {
+namespace GrpcCommon {
 
 // TODO(mattklein123): Stats
 
@@ -127,44 +127,7 @@ private:
   const LocalInfo::LocalInfo& local_info_;
 };
 
-/**
- * Access log Instance that streams HTTP logs over gRPC.
- */
-class HttpGrpcAccessLog : public Common::ImplBase {
-public:
-  HttpGrpcAccessLog(AccessLog::FilterPtr&& filter,
-                    envoy::config::accesslog::v2::HttpGrpcAccessLogConfig config,
-                    ThreadLocal::SlotAllocator& tls,
-                    GrpcAccessLoggerCacheSharedPtr access_logger_cache);
-
-  static void responseFlagsToAccessLogResponseFlags(
-      envoy::data::accesslog::v2::AccessLogCommon& common_access_log,
-      const StreamInfo::StreamInfo& stream_info);
-
-private:
-  /**
-   * Per-thread cached logger.
-   */
-  struct ThreadLocalLogger : public ThreadLocal::ThreadLocalObject {
-    ThreadLocalLogger(GrpcAccessLoggerSharedPtr logger);
-
-    const GrpcAccessLoggerSharedPtr logger_;
-  };
-
-  // Common::ImplBase
-  void emitLog(const Http::HeaderMap& request_headers, const Http::HeaderMap& response_headers,
-               const Http::HeaderMap& response_trailers,
-               const StreamInfo::StreamInfo& stream_info) override;
-
-  const envoy::config::accesslog::v2::HttpGrpcAccessLogConfig config_;
-  const ThreadLocal::SlotPtr tls_slot_;
-  const GrpcAccessLoggerCacheSharedPtr access_logger_cache_;
-  std::vector<Http::LowerCaseString> request_headers_to_log_;
-  std::vector<Http::LowerCaseString> response_headers_to_log_;
-  std::vector<Http::LowerCaseString> response_trailers_to_log_;
-};
-
-} // namespace HttpGrpc
+} // namespace GrpcCommon
 } // namespace AccessLoggers
 } // namespace Extensions
 } // namespace Envoy
