+++ test/extensions/access_loggers/grpc/http_grpc_access_log_impl_test.cc
@@ -3,7 +3,7 @@
 #include "common/buffer/zero_copy_input_stream_impl.h"
 #include "common/network/address_impl.h"
 
-#include "extensions/access_loggers/http_grpc/grpc_access_log_impl.h"
+#include "extensions/access_loggers/grpc/http_grpc_access_log_impl.h"
 
 #include "test/mocks/access_log/mocks.h"
 #include "test/mocks/grpc/mocks.h"
@@ -25,281 +25,17 @@ namespace AccessLoggers {
 namespace HttpGrpc {
 namespace {
 
-constexpr std::chrono::milliseconds FlushInterval(10);
-
-class GrpcAccessLoggerImplTest : public testing::Test {
-public:
-  using MockAccessLogStream = Grpc::MockAsyncStream;
-  using AccessLogCallbacks =
-      Grpc::AsyncStreamCallbacks<envoy::service::accesslog::v2::StreamAccessLogsResponse>;
-
-  void initLogger(std::chrono::milliseconds buffer_flush_interval_msec, size_t buffer_size_bytes) {
-    timer_ = new Event::MockTimer(&dispatcher_);
-    EXPECT_CALL(*timer_, enableTimer(buffer_flush_interval_msec));
-    logger_ = std::make_unique<GrpcAccessLoggerImpl>(Grpc::RawAsyncClientPtr{async_client_},
-                                                     log_name_, buffer_flush_interval_msec,
-                                                     buffer_size_bytes, dispatcher_, local_info_);
-  }
-
-  void expectStreamStart(MockAccessLogStream& stream, AccessLogCallbacks** callbacks_to_set) {
-    EXPECT_CALL(*async_client_, startRaw(_, _, _))
-        .WillOnce(Invoke([&stream, callbacks_to_set](absl::string_view, absl::string_view,
-                                                     Grpc::RawAsyncStreamCallbacks& callbacks) {
-          *callbacks_to_set = dynamic_cast<AccessLogCallbacks*>(&callbacks);
-          return &stream;
-        }));
-  }
-
-  void expectStreamMessage(MockAccessLogStream& stream, const std::string& expected_message_yaml) {
-    envoy::service::accesslog::v2::StreamAccessLogsMessage expected_message;
-    TestUtility::loadFromYaml(expected_message_yaml, expected_message);
-    EXPECT_CALL(stream, sendMessageRaw_(_, false))
-        .WillOnce(Invoke([expected_message](Buffer::InstancePtr& request, bool) {
-          envoy::service::accesslog::v2::StreamAccessLogsMessage message;
-          Buffer::ZeroCopyInputStreamImpl request_stream(std::move(request));
-          EXPECT_TRUE(message.ParseFromZeroCopyStream(&request_stream));
-          EXPECT_EQ(message.DebugString(), expected_message.DebugString());
-        }));
-  }
-
-  std::string log_name_ = "test_log_name";
-  LocalInfo::MockLocalInfo local_info_;
-  Event::MockTimer* timer_ = nullptr;
-  Event::MockDispatcher dispatcher_;
-  Grpc::MockAsyncClient* async_client_{new Grpc::MockAsyncClient};
-  std::unique_ptr<GrpcAccessLoggerImpl> logger_;
-};
-
-// Test basic stream logging flow.
-TEST_F(GrpcAccessLoggerImplTest, BasicFlow) {
-  InSequence s;
-  initLogger(FlushInterval, 0);
-
-  // Start a stream for the first log.
-  MockAccessLogStream stream;
-  AccessLogCallbacks* callbacks;
-  expectStreamStart(stream, &callbacks);
-  EXPECT_CALL(local_info_, node());
-  expectStreamMessage(stream, R"EOF(
-identifier:
-  node:
-    id: node_name
-    cluster: cluster_name
-    locality:
-      zone: zone_name
-  log_name: test_log_name
-http_logs:
-  log_entry:
-    request:
-      path: /test/path1
-)EOF");
-  envoy::data::accesslog::v2::HTTPAccessLogEntry entry;
-  entry.mutable_request()->set_path("/test/path1");
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-
-  expectStreamMessage(stream, R"EOF(
-http_logs:
-  log_entry:
-    request:
-      path: /test/path2
-)EOF");
-  entry.mutable_request()->set_path("/test/path2");
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-
-  // Verify that sending an empty response message doesn't do anything bad.
-  callbacks->onReceiveMessage(
-      std::make_unique<envoy::service::accesslog::v2::StreamAccessLogsResponse>());
-
-  // Close the stream and make sure we make a new one.
-  callbacks->onRemoteClose(Grpc::Status::Internal, "bad");
-  expectStreamStart(stream, &callbacks);
-  EXPECT_CALL(local_info_, node());
-  expectStreamMessage(stream, R"EOF(
-identifier:
-  node:
-    id: node_name
-    cluster: cluster_name
-    locality:
-      zone: zone_name
-  log_name: test_log_name
-http_logs:
-  log_entry:
-    request:
-      path: /test/path3
-)EOF");
-  entry.mutable_request()->set_path("/test/path3");
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-}
-
-// Test that stream failure is handled correctly.
-TEST_F(GrpcAccessLoggerImplTest, StreamFailure) {
-  InSequence s;
-  initLogger(FlushInterval, 0);
-
-  EXPECT_CALL(*async_client_, startRaw(_, _, _))
-      .WillOnce(Invoke(
-          [](absl::string_view, absl::string_view, Grpc::RawAsyncStreamCallbacks& callbacks) {
-            callbacks.onRemoteClose(Grpc::Status::Internal, "bad");
-            return nullptr;
-          }));
-  EXPECT_CALL(local_info_, node());
-  envoy::data::accesslog::v2::HTTPAccessLogEntry entry;
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-}
-
-// Test that log entries are batched.
-TEST_F(GrpcAccessLoggerImplTest, Batching) {
-  InSequence s;
-  initLogger(FlushInterval, 100);
-
-  MockAccessLogStream stream;
-  AccessLogCallbacks* callbacks;
-  expectStreamStart(stream, &callbacks);
-  EXPECT_CALL(local_info_, node());
-  const std::string path1(30, '1');
-  const std::string path2(30, '2');
-  const std::string path3(80, '3');
-  expectStreamMessage(stream, fmt::format(R"EOF(
-identifier:
-  node:
-    id: node_name
-    cluster: cluster_name
-    locality:
-      zone: zone_name
-  log_name: test_log_name
-http_logs:
-  log_entry:
-  - request:
-      path: "{}"
-  - request:
-      path: "{}"
-  - request:
-      path: "{}"
-)EOF",
-                                          path1, path2, path3));
-  envoy::data::accesslog::v2::HTTPAccessLogEntry entry;
-  entry.mutable_request()->set_path(path1);
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-  entry.mutable_request()->set_path(path2);
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-  entry.mutable_request()->set_path(path3);
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-
-  const std::string path4(120, '4');
-  expectStreamMessage(stream, fmt::format(R"EOF(
-http_logs:
-  log_entry:
-    request:
-      path: "{}"
-)EOF",
-                                          path4));
-  entry.mutable_request()->set_path(path4);
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-}
-
-// Test that log entries are flushed periodically.
-TEST_F(GrpcAccessLoggerImplTest, Flushing) {
-  InSequence s;
-  initLogger(FlushInterval, 100);
-
-  // Nothing to do yet.
-  EXPECT_CALL(*timer_, enableTimer(FlushInterval));
-  timer_->invokeCallback();
-
-  envoy::data::accesslog::v2::HTTPAccessLogEntry entry;
-  // Not enough data yet to trigger flush on batch size.
-  entry.mutable_request()->set_path("/test/path1");
-  logger_->log(envoy::data::accesslog::v2::HTTPAccessLogEntry(entry));
-
-  MockAccessLogStream stream;
-  AccessLogCallbacks* callbacks;
-  expectStreamStart(stream, &callbacks);
-  EXPECT_CALL(local_info_, node());
-  expectStreamMessage(stream, fmt::format(R"EOF(
-  identifier:
-    node:
-      id: node_name
-      cluster: cluster_name
-      locality:
-        zone: zone_name
-    log_name: test_log_name
-  http_logs:
-    log_entry:
-    - request:
-        path: /test/path1
-  )EOF"));
-  EXPECT_CALL(*timer_, enableTimer(FlushInterval));
-  timer_->invokeCallback();
-
-  // Flush on empty message does nothing.
-  EXPECT_CALL(*timer_, enableTimer(FlushInterval));
-  timer_->invokeCallback();
-}
-
-class GrpcAccessLoggerCacheImplTest : public testing::Test {
-public:
-  GrpcAccessLoggerCacheImplTest() {
-    logger_cache_ = std::make_unique<GrpcAccessLoggerCacheImpl>(async_client_manager_, scope_, tls_,
-                                                                local_info_);
-  }
-
-  void expectClientCreation() {
-    factory_ = new Grpc::MockAsyncClientFactory;
-    async_client_ = new Grpc::MockAsyncClient;
-    EXPECT_CALL(async_client_manager_, factoryForGrpcService(_, _, false))
-        .WillOnce(Invoke([this](const envoy::api::v2::core::GrpcService&, Stats::Scope&, bool) {
-          EXPECT_CALL(*factory_, create()).WillOnce(Invoke([this] {
-            return Grpc::RawAsyncClientPtr{async_client_};
-          }));
-          return Grpc::AsyncClientFactoryPtr{factory_};
-        }));
-  }
-
-  LocalInfo::MockLocalInfo local_info_;
-  NiceMock<ThreadLocal::MockInstance> tls_;
-  Grpc::MockAsyncClientManager async_client_manager_;
-  Grpc::MockAsyncClient* async_client_ = nullptr;
-  Grpc::MockAsyncClientFactory* factory_ = nullptr;
-  std::unique_ptr<GrpcAccessLoggerCacheImpl> logger_cache_;
-  NiceMock<Stats::MockIsolatedStatsStore> scope_;
-};
-
-TEST_F(GrpcAccessLoggerCacheImplTest, Deduplication) {
-  InSequence s;
-
-  ::envoy::config::accesslog::v2::CommonGrpcAccessLogConfig config;
-  config.set_log_name("log-1");
-  config.mutable_grpc_service()->mutable_envoy_grpc()->set_cluster_name("cluster-1");
-
-  expectClientCreation();
-  GrpcAccessLoggerSharedPtr logger1 = logger_cache_->getOrCreateLogger(config);
-  EXPECT_EQ(logger1, logger_cache_->getOrCreateLogger(config));
-
-  // Changing log name leads to another logger.
-  config.set_log_name("log-2");
-  expectClientCreation();
-  EXPECT_NE(logger1, logger_cache_->getOrCreateLogger(config));
-
-  config.set_log_name("log-1");
-  EXPECT_EQ(logger1, logger_cache_->getOrCreateLogger(config));
-
-  // Changing cluster name leads to another logger.
-  config.mutable_grpc_service()->mutable_envoy_grpc()->set_cluster_name("cluster-2");
-  expectClientCreation();
-  EXPECT_NE(logger1, logger_cache_->getOrCreateLogger(config));
-}
-
-class MockGrpcAccessLogger : public GrpcAccessLogger {
+class MockGrpcAccessLogger : public GrpcCommon::GrpcAccessLogger {
 public:
   // GrpcAccessLogger
   MOCK_METHOD1(log, void(envoy::data::accesslog::v2::HTTPAccessLogEntry&& entry));
 };
 
-class MockGrpcAccessLoggerCache : public GrpcAccessLoggerCache {
+class MockGrpcAccessLoggerCache : public GrpcCommon::GrpcAccessLoggerCache {
 public:
   // GrpcAccessLoggerCache
   MOCK_METHOD1(getOrCreateLogger,
-               GrpcAccessLoggerSharedPtr(
+               GrpcCommon::GrpcAccessLoggerSharedPtr(
                    const ::envoy::config::accesslog::v2::CommonGrpcAccessLogConfig& config));
 };
 
@@ -861,37 +597,6 @@ response:
 )EOF");
     access_log_->log(&request_headers, &response_headers, &response_trailers, stream_info);
   }
-}
-
-TEST(responseFlagsToAccessLogResponseFlagsTest, All) {
-  NiceMock<StreamInfo::MockStreamInfo> stream_info;
-  ON_CALL(stream_info, hasResponseFlag(_)).WillByDefault(Return(true));
-  envoy::data::accesslog::v2::AccessLogCommon common_access_log;
-  HttpGrpcAccessLog::responseFlagsToAccessLogResponseFlags(common_access_log, stream_info);
-
-  envoy::data::accesslog::v2::AccessLogCommon common_access_log_expected;
-  common_access_log_expected.mutable_response_flags()->set_failed_local_healthcheck(true);
-  common_access_log_expected.mutable_response_flags()->set_no_healthy_upstream(true);
-  common_access_log_expected.mutable_response_flags()->set_upstream_request_timeout(true);
-  common_access_log_expected.mutable_response_flags()->set_local_reset(true);
-  common_access_log_expected.mutable_response_flags()->set_upstream_remote_reset(true);
-  common_access_log_expected.mutable_response_flags()->set_upstream_connection_failure(true);
-  common_access_log_expected.mutable_response_flags()->set_upstream_connection_termination(true);
-  common_access_log_expected.mutable_response_flags()->set_upstream_overflow(true);
-  common_access_log_expected.mutable_response_flags()->set_no_route_found(true);
-  common_access_log_expected.mutable_response_flags()->set_delay_injected(true);
-  common_access_log_expected.mutable_response_flags()->set_fault_injected(true);
-  common_access_log_expected.mutable_response_flags()->set_rate_limited(true);
-  common_access_log_expected.mutable_response_flags()->mutable_unauthorized_details()->set_reason(
-      envoy::data::accesslog::v2::ResponseFlags_Unauthorized_Reason::
-          ResponseFlags_Unauthorized_Reason_EXTERNAL_SERVICE);
-  common_access_log_expected.mutable_response_flags()->set_rate_limit_service_error(true);
-  common_access_log_expected.mutable_response_flags()->set_downstream_connection_termination(true);
-  common_access_log_expected.mutable_response_flags()->set_upstream_retry_limit_exceeded(true);
-  common_access_log_expected.mutable_response_flags()->set_stream_idle_timeout(true);
-  common_access_log_expected.mutable_response_flags()->set_invalid_envoy_request_headers(true);
-
-  EXPECT_EQ(common_access_log_expected.DebugString(), common_access_log.DebugString());
 }
 
 TEST_F(HttpGrpcAccessLogTest, LogWithRequestMethod) {
